requirements.txt
Flask
Flask-SocketIO
opencv-python
mediapipe
tensorflow
numpyscikit-learn
tqdm
matplotlib